+++ src/main/java/com/example/demo/infrastructure/repository/JpaProductPrices.java
@@ -14,7 +14,7 @@ public class JpaProductPrices implements ProductPricesRepository{
     private EntityManager entityManager;
 
     @Override
-    public List<ProductPrices> findProductByPrice(Long brand_id, Long product_id, LocalDateTime applicationDate) {
+    public List<ProductPrices> findProductByPrice(Long brand_id, int product_id, LocalDateTime applicationDate) {
             String jpql = "SELECT p FROM ProductPrices p WHERE p.brand_id = :brand_id " +
                     "AND p.product_id = :product_id " +
                     "AND :applicationDate BETWEEN p.start_date AND p.end_date " +
